refactor: organize filter sort query into folders
