+++ clients/java/src/main/java/io/camunda/zeebe/client/api/search/query/UserTaskQuery.java
@@ -13,25 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/*
- * Copyright © 2017 camunda services GmbH (info@camunda.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.camunda.zeebe.client.api.search;
 
+package io.camunda.zeebe.client.api.search.query;
+
+import io.camunda.zeebe.client.api.search.FinalSearchQueryStep;
+import io.camunda.zeebe.client.api.search.TypedSearchQueryRequest;
+import io.camunda.zeebe.client.api.search.filter.UserTaskFilter;
 import io.camunda.zeebe.client.api.search.response.UserTask;
+import io.camunda.zeebe.client.api.search.sort.UserTaskSort;
 
 public interface UserTaskQuery
     extends TypedSearchQueryRequest<UserTaskFilter, UserTaskSort, UserTaskQuery>,
-        FinalSearchQueryStep<UserTask> {}
+    FinalSearchQueryStep<UserTask> {}
