+++ clients/java/src/main/java/io/camunda/zeebe/client/api/search/sort/DecisionDefinitionSort.java
@@ -1,0 +1,42 @@
+/*
+ * Copyright © 2017 camunda services GmbH (info@camunda.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.camunda.zeebe.client.api.search.sort;
+
+import io.camunda.zeebe.client.api.search.TypedSearchQueryRequest.SearchRequestSort;
+
+public interface DecisionDefinitionSort extends SearchRequestSort<DecisionDefinitionSort> {
+
+  /** Sort by decision key. */
+  DecisionDefinitionSort decisionKey();
+
+  /** Sort by dmn decision id. */
+  DecisionDefinitionSort dmnDecisionId();
+
+  /** Sort by dmn decision name. */
+  DecisionDefinitionSort dmnDecisionName();
+
+  /** Sort by version. */
+  DecisionDefinitionSort version();
+
+  /** Sort by dmn decision requirements id. */
+  DecisionDefinitionSort dmnDecisionRequirementsId();
+
+  /** Sort by decision requirements key. */
+  DecisionDefinitionSort decisionRequirementsKey();
+
+  /** Sort by tenant id. */
+  DecisionDefinitionSort tenantId();
+}
