+++ clients/java/src/main/java/io/camunda/zeebe/client/impl/search/query/DecisionRequirementsQueryImpl.java
@@ -13,30 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/*
- * Copyright © 2017 camunda services GmbH (info@camunda.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.camunda.zeebe.client.impl.search;
+
+package io.camunda.zeebe.client.impl.search.query;
 
 import static io.camunda.zeebe.client.api.search.SearchRequestBuilders.decisionRequirementsSort;
 import static io.camunda.zeebe.client.api.search.SearchRequestBuilders.searchRequestPage;
 
 import io.camunda.zeebe.client.api.JsonMapper;
-import io.camunda.zeebe.client.api.search.DecisionRequirementsFilter;
-import io.camunda.zeebe.client.api.search.DecisionRequirementsQuery;
-import io.camunda.zeebe.client.api.search.DecisionRequirementsSort;
+import io.camunda.zeebe.client.api.search.filter.DecisionRequirementsFilter;
+import io.camunda.zeebe.client.api.search.query.DecisionRequirementsQuery;
+import io.camunda.zeebe.client.api.search.sort.DecisionRequirementsSort;
 import io.camunda.zeebe.client.api.search.FinalSearchQueryStep;
 import io.camunda.zeebe.client.api.search.SearchRequestBuilders;
 import io.camunda.zeebe.client.api.search.SearchRequestPage;
@@ -44,6 +30,10 @@ import io.camunda.zeebe.client.api.search.response.DecisionRequirements;
 import io.camunda.zeebe.client.api.search.response.SearchQueryResponse;
 import io.camunda.zeebe.client.impl.http.HttpClient;
 import io.camunda.zeebe.client.impl.http.HttpZeebeFuture;
+import io.camunda.zeebe.client.impl.search.SearchRequestPageImpl;
+import io.camunda.zeebe.client.impl.search.SearchResponseMapper;
+import io.camunda.zeebe.client.impl.search.TypedSearchRequestPropertyProvider;
+import io.camunda.zeebe.client.impl.search.sort.DecisionRequirementsSortImpl;
 import io.camunda.zeebe.client.protocol.rest.DecisionRequirementsFilterRequest;
 import io.camunda.zeebe.client.protocol.rest.DecisionRequirementsSearchQueryRequest;
 import io.camunda.zeebe.client.protocol.rest.DecisionRequirementsSearchQueryResponse;
