+++ clients/java/src/main/java/io/camunda/zeebe/client/impl/search/query/UserTaskQueryImpl.java
@@ -13,22 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-/*
- * Copyright © 2017 camunda services GmbH (info@camunda.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.camunda.zeebe.client.impl.search;
+
+package io.camunda.zeebe.client.impl.search.query;
 
 import static io.camunda.zeebe.client.api.search.SearchRequestBuilders.searchRequestPage;
 import static io.camunda.zeebe.client.api.search.SearchRequestBuilders.userTaskFilter;
@@ -37,13 +23,17 @@ import static io.camunda.zeebe.client.api.search.SearchRequestBuilders.userTaskS
 import io.camunda.zeebe.client.api.JsonMapper;
 import io.camunda.zeebe.client.api.search.FinalSearchQueryStep;
 import io.camunda.zeebe.client.api.search.SearchRequestPage;
-import io.camunda.zeebe.client.api.search.UserTaskFilter;
-import io.camunda.zeebe.client.api.search.UserTaskQuery;
-import io.camunda.zeebe.client.api.search.UserTaskSort;
+import io.camunda.zeebe.client.api.search.filter.UserTaskFilter;
+import io.camunda.zeebe.client.api.search.query.UserTaskQuery;
+import io.camunda.zeebe.client.api.search.sort.UserTaskSort;
 import io.camunda.zeebe.client.api.search.response.SearchQueryResponse;
 import io.camunda.zeebe.client.api.search.response.UserTask;
 import io.camunda.zeebe.client.impl.http.HttpClient;
 import io.camunda.zeebe.client.impl.http.HttpZeebeFuture;
+import io.camunda.zeebe.client.impl.search.SearchRequestPageImpl;
+import io.camunda.zeebe.client.impl.search.SearchResponseMapper;
+import io.camunda.zeebe.client.impl.search.TypedSearchRequestPropertyProvider;
+import io.camunda.zeebe.client.impl.search.sort.UserTaskSortImpl;
 import io.camunda.zeebe.client.protocol.rest.UserTaskFilterRequest;
 import io.camunda.zeebe.client.protocol.rest.UserTaskSearchQueryRequest;
 import io.camunda.zeebe.client.protocol.rest.UserTaskSearchQueryResponse;
