+++ clients/java/src/main/java/io/camunda/zeebe/client/impl/search/sort/DecisionDefinitionSortImpl.java
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.zeebe.client.impl.search;
+package io.camunda.zeebe.client.impl.search.sort;
 
-import io.camunda.zeebe.client.api.search.DecisionDefinitionSort;
+import io.camunda.zeebe.client.api.search.sort.DecisionDefinitionSort;
+import io.camunda.zeebe.client.impl.search.SearchQuerySortBase;
 
 public class DecisionDefinitionSortImpl extends SearchQuerySortBase<DecisionDefinitionSort>
     implements DecisionDefinitionSort {
