+++ clients/java/src/main/java/io/camunda/zeebe/client/impl/search/sort/DecisionRequirementsSortImpl.java
@@ -13,9 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.zeebe.client.impl.search;
+package io.camunda.zeebe.client.impl.search.sort;
 
-import io.camunda.zeebe.client.api.search.DecisionRequirementsSort;
+import io.camunda.zeebe.client.api.search.sort.DecisionRequirementsSort;
+import io.camunda.zeebe.client.impl.search.SearchQuerySortBase;
 
 public class DecisionRequirementsSortImpl extends SearchQuerySortBase<DecisionRequirementsSort>
     implements DecisionRequirementsSort {
